fix: переносит FavoriteItem в компоненты
